changed components names and folder for them
