--- src/components/index/header/Header.jsx
@@ -1,28 +1,0 @@
-import React from "react";
-
-const Header = () => {
-  return (
-    <div className="container">
-      <div className="header__wrapper">
-        <div className="header__left">
-          <a className="header__logo-link header__logo-link--active">
-            <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-          </a>
-        </div>
-        <nav className="header__nav">
-          <ul className="header__nav-list">
-            <li className="header__nav-item user">
-              <a className="header__nav-link header__nav-link--profile" href="#">
-                <div className="header__avatar-wrapper user__avatar-wrapper">
-                </div>
-                <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-              </a>
-            </li>
-          </ul>
-        </nav>
-      </div>
-    </div>
-  );
-};
-
-export default Header;
