+++ src/components/places/places.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Article from "../article/Article";
+import Article from "../article/article";
 import PropTypes from "prop-types";
 
 const Places = (props) => {
@@ -11,9 +11,9 @@ const Places = (props) => {
       <form className="places__sorting" action="#" method="get">
         <span className="places__sorting-caption">Sort by</span>
         <span className="places__sorting-type" tabIndex="0">
-                      Popular
+          Popular
           <svg className="places__sorting-arrow" width="7" height="4">
-            {/* <use xlink:href="#icon-arrow-select"></use> */}
+            <use xlinkHref="#icon-arrow-select"></use>
           </svg>
         </span>
         <ul className="places__options places__options--custom places__options--opened">
@@ -23,11 +23,11 @@ const Places = (props) => {
           <li className="places__option" tabIndex="0">Top rated first</li>
         </ul>
         {/* <select className="places__sorting-type" id="places-sorting">
-                      <option className="places__option" value="popular" selected="">Popular</option>
-                      <option className="places__option" value="to-high">Price: low to high</option>
-                      <option className="places__option" value="to-low">Price: high to low</option>
-                      <option className="places__option" value="top-rated">Top rated first</option>
-                    </select> */}
+          <option className="places__option" value="popular" selected="">Popular</option>
+          <option className="places__option" value="to-high">Price: low to high</option>
+          <option className="places__option" value="to-low">Price: high to low</option>
+          <option className="places__option" value="top-rated">Top rated first</option>
+        </select> */}
       </form>
       <div className="cities__places-list places__list tabs__content">
         <Article offersCount = {offersCount} />
